Cleaned up and added configurability to brightest object plugin
