+++ src/mousetrap/mousetrap.yaml
@@ -2,12 +2,11 @@
 #            To enable a plugin, list it here.
 assembly:
 - mousetrap.plugins.camera.CameraPlugin
-#- mousetrap.plugins.display.DisplayPlugin
-#- mousetrap.plugins.nose.NoseJoystickPlugin
-#- mousetrap.plugins.eyes.EyesPlugin
-#- mousetrap.vision.FeatureDetectorClearCachePlugin
-#- mousetrap.plugins.bright_object.Bright
-- mousetrap.plugins.bright_object_joystick.Bright
+- mousetrap.plugins.display.DisplayPlugin
+- mousetrap.plugins.nose.NoseJoystickPlugin
+- mousetrap.plugins.eyes.EyesPlugin
+- mousetrap.vision.FeatureDetectorClearCachePlugin
+#- mousetrap.plugins.bright.Bright
 
 
 # camera - Configuration for the built-in camera.
@@ -53,6 +52,9 @@ classes:
       scale_factor: 1.1
   mousetrap.plugins.nose.NoseJoystickPlugin:
     threshold: 5
+  mousetrap.plugins.bright.Bright:
+    dead_zone: 50
+    
 
 # haar_files - A mapping of haar cascade files. Relative paths are relative
 #              to the mousetrap package directory. Plugins, if they come with
